+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Moved functionality into `sys` module
 - Introduced `MatrixStack` type
+- Introduced `Shader` and `Program` types
 
 
 0.1.0